moved test code from integration test to the chainflip.rs to minimize impact and make things easier to read
