+++ state-chain/cf-integration-tests/Cargo.toml
@@ -22,7 +22,6 @@ log = { version = '0.4.16', default-features = false }
 libsecp256k1 = { version = "0.7", features = ['static-context'] }
 rand = "0.8.4"
 hex-literal = "0.4.1"
-hex = '0.4.3'
 secp256k1 = { version = "0.27", features = ["rand-std"] }
 arrayref = '0.3.6'
