+++ state-chain/cf-integration-tests/src/lib.rs
@@ -9,7 +9,6 @@ mod threshold_signing;
 
 mod account;
 mod authorities;
-mod events;
 mod funding;
 mod genesis;
 mod governance;
